Lint for Linux from macOS, and fix what that found
CI failed on a finding this machine could not see: a redundant int64
conversion in a file guarded by //go:build linux. A linter running under
one GOOS never compiles the files belonging to another, so the whole
platform-specific half of the tree was outside the local loop.

make lint now runs golangci-lint three times -- for this host, and for
linux/amd64 and linux/arm64. It costs about two seconds and closes a gap
that otherwise costs a push, a wait, and a read of the CI log.

The conversion itself is gone rather than annotated: comparing against
untyped constants works whatever width Statfs_t.Type has on a given
architecture, which is what made the conversion redundant on ours and
necessary-looking in the first place.

+++ internal/config/fstype_linux.go
@@ -23,7 +23,10 @@ const (
 // and most of what runs on it is local. Calling every FUSE mount unsafe would
 // make the check noisy enough to be turned off.
 func fsTypeName(st *unix.Statfs_t) string {
-	switch int64(st.Type) {
+	// Compared against untyped constants rather than converted: Statfs_t.Type
+	// is int64 on amd64 and arm64 but not on every architecture, and a
+	// conversion that is redundant on one is a lint failure there.
+	switch st.Type {
 	case nfsSuperMagic:
 		return "nfs"
 	case smbSuperMagic:
